refactor: 테스트 클래스 분리, 테스트 메서드 수정
