+++ src/test/java/com/mission2/racingcar/CarRacingServiceTests.java
@@ -1,6 +1,6 @@
 package com.mission2.racingcar;
 
-import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -9,13 +9,12 @@ import java.util.*;
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 
-public class CarRacingTests {
+public class CarRacingServiceTests {
 
-    static CarRacingService service;
-    static Car car = new Car("TestCar", 0);
+    CarRacingService service;
 
-    @BeforeAll
-    public static void beforeAll() {
+    @BeforeEach
+    public void beforeEach() {
         service = new CarRacingService();
     }
 
@@ -33,8 +32,8 @@ public class CarRacingTests {
     }
 
     @Test
-    @DisplayName("자동차 이름은 문자열 5자 이하이다")
-    void 자동차_이름_체크() {
+    @DisplayName("자동차 이름 5자 초과는 허용되지 않는다")
+    void 자동차_이름_길이_체크() {
         //given
         String input = "AAA,BBBBBB,CCC,DDD,EEE"; // Scanner 입력
 
@@ -60,32 +59,6 @@ public class CarRacingTests {
     }
 
     @Test
-    @DisplayName("Random 숫자의 값이 0~3 이면 멈춘다")
-    void Random_값_정지() {
-        //given
-        int stopValues = 3;
-
-        //when
-        int stop = car.stopOrGoByRandomNumber(stopValues);
-
-        //then
-        assertThat(stop).isZero();
-    }
-
-    @Test
-    @DisplayName("Random 숫자의 값이 4~9 이면 전진한다")
-    void Random_값_전진() {
-        //given
-        int forwardValues = 4;
-
-        //when
-        int forward = car.stopOrGoByRandomNumber(forwardValues);
-
-        //then
-        assertThat(forward).isEqualTo(1);
-    }
-
-    @Test
     @DisplayName("사용자가 입력한 시도 횟수만큼 반복한다")
     void 시도_횟수만큼_반복() {
         //given
@@ -99,22 +72,6 @@ public class CarRacingTests {
 
         //then
         assertThat(race.getGameCount()).isEqualTo(gameCount);
-    }
-
-    @Test
-    @DisplayName("종료시 우승자를 출력한다")
-    void 종료시_우승자_출력() {
-        //given
-        Car[] carList = { new Car("AAA", 3), new Car("BBB", 3),
-                new Car("CCC", 4), new Car("DDD", 5) };
-        Race race = new Race(5, Arrays.asList(carList));
-        String[] result = race.getWinners();
-
-        //when
-
-        //then
-        assertThat(result).hasSize(1);
-        assertThat(result[0]).isEqualTo("DDD");
     }
 
 }
